refactor useHours
