+++ pages/welcome.tsx
@@ -1,7 +1,6 @@
-import { Box, Button, Container, Stack, TextField, Typography } from "@mui/material";
+import { Button, Container, Stack, TextField, Typography } from "@mui/material";
 import "react-datepicker/dist/react-datepicker.css";
 import { Layout } from "../src/components/layout";
-import { ContentHeader } from "../src/components/content-header";
 import { useRouter } from "next/router";
 import { GetServerSideProps } from "next";
 import { withServerSideSession } from "../src/server/with-session";
