--- src/hooks/use-hours.ts
@@ -1,26 +1,0 @@
-import { getAxios } from "../get-axios";
-import { useCallback, useState } from "react";
-import { GetHoursHandlerResponse, ProjectHours } from "../../pages/api/user/hours";
-
-export const useHours = () => {
-    const [ isLoading, setIsLoading ] = useState(false);
-    const [ hours, setHours ] = useState<ProjectHours[] | null>(null);
-
-    const load = useCallback((from: string, to: string, uid: string = '', projectId?: number) => {
-        setIsLoading(true);
-        getAxios().get<GetHoursHandlerResponse>(`/user/hours?from=${ from }&to=${ to }&uid=${ uid }&project_id=${ projectId }`)
-            .then(({ data }) => {
-                setHours(data)
-                setHours(data)
-            })
-            .finally(() => {
-                setIsLoading(false);
-            });
-    }, []);
-
-    return {
-        load,
-        isLoading,
-        hours,
-    }
-}
